Добавил join-ы авторов и жанров в запросы по книгам; удалил избыточные ошибки Dao; добавил ошибку удаления книги, если задан отсутствующий в базе идентификатор; добавил юнит-тест на ошибку удаления

+++ mybooks/src/main/java/ru/otus/mybooks/service/BookServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import ru.otus.mybooks.dao.BookDao;
 import ru.otus.mybooks.domain.Book;
 import ru.otus.mybooks.dto.BookDto;
+import ru.otus.mybooks.exception.BookServiceRemoveBookException;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -27,12 +28,14 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public BookDto saveBook(Book book) {
-        return new BookDto(dao.find(book).orElseGet(() -> dao.update(book)));
+    public void saveBook(Book book) {
+        if (dao.find(book).isEmpty()) dao.update(book);
     }
 
     @Override
     public void removeBook(long bookNum) {
+        if (dao.getById(bookNum).isEmpty())
+            throw new BookServiceRemoveBookException(String.format("The book with ID %d was not found in the database.", bookNum));
         dao.deleteById(bookNum);
     }
 }
